Split out availability / distribution queries
