+++ tests/conftest.py
@@ -5,11 +5,11 @@ from hutch_bunny.core.rquest_models.cohort import Cohort
 from hutch_bunny.core.rquest_models.group import Group
 from hutch_bunny.core.rquest_models.rule import Rule
 from hutch_bunny.core.db_manager import SyncDBManager
-from hutch_bunny.core.rquest_models.query import (
-    AvailabilityQuery,
+from hutch_bunny.core.rquest_models.distribution import (
     DistributionQuery,
     DistributionQueryType,
 )
+from hutch_bunny.core.rquest_models.availability import AvailabilityQuery
 from hutch_bunny.core.settings import Settings
 import hutch_bunny.core.db as db
